Support handling nested classes
Without understanding the module hierarchy, we can't distinguish
between a module and a class in a dotted name like `foo.bar.A.B`, and
will insert code like `from foo.bar.A import B`, even though `A` is a
class.

Work around this with a cruddy heuristic that names beginning with a
capital letter are classes and not modules.

Another approach would be to introduce alternate punctuation like ! or
: to separate the module and class part of a name. This would have the
disadvantage of being pretty nonstandard but also not being an awful
hack.